requirements.txt
openai>=1.40.0
requests>=2.31.0
PyYAML>=6.0.1
ddgs>=6.16.1
wikipedia>=1.4.0
diffusers>=0.30.0
transformers>=4.44.0
accelerate>=0.33.0
safetensors>=0.4.3
Pillow>=10.4.0
